refactor: decommission ss_provision.c/h
ss_provision.c/.h doesn't hold provisioning, only
generate_dir_table_from_blob after recent refactoring and onomondo-uicc
introduction.
